// planning-idsdp-pps-frontend/src/app/main/modules/project-concept-management/models/mode-of-finance.model.ts
export class ModeOfFinanceModel {
    modeId: number;
    totalAmount: number;
    gobAmount: number;
    feGobAmount: number;
    ownFundAmount: number;
    feOwnFundAmount: number;
    paAmount: number;
    rpaAmount: number;
    dpaAmount: number;
    otherAmount: number;
    feOtherAmount: number;
    projectConceptMasterId: number;
    id: number;
    uuid: string;
}